% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundprog_one.R
\name{soundprog_one}
\alias{soundprog_one}
\title{Calculate Noise Propagation Patterns}
\usage{
soundprog_one(
  sgrid,
  src_one,
  data_one,
  atmos,
  geo,
  new_barrier = TRUE,
  all_basins = NULL,
  source_offset = 0.34,
  receiver_offset = 1
)
}
\arguments{
\item{sgrid}{A spatial grid configuration representing the domain of analysis.}

\item{src_one}{A source of sound, either as a point or polygon (sf object).}

\item{data_one}{A data.frame containing sound frequency (\code{freq}),
sound pressure levels (\code{db}), and measurement distance (\code{dist}).}

\item{atmos}{Atmospheric conditions including:
\itemize{
  \item \code{temp}: Temperature in degrees Celsius (°C).
  \item \code{hum}: Relative humidity as a percentage (%).
  \item \code{ws}: Wind speed in meters per second (m/s).
  \item \code{wd}: Wind direction in degrees.
  \item \code{seas_cond}: Seasonal conditions, where:
    \enumerate{
      \item Clear, calm summer day.
      \item Clear, calm winter day.
      \item Clear, calm summer night.
      \item Clear, calm winter night.
      \item Clear, windy summer day.
      \item Clear, windy winter day.
      \item Clear, windy summer night.
      \item Clear, windy winter night.
      \item Cloudy, calm.
      \item Cloudy, windy.
    }
}}

\item{geo}{Geographic data, including:
\itemize{
  \item \code{dem}: Digital Elevation Model (DEM) raster.
  \item \code{land}: Land cover data.
}}

\item{new_barrier}{Logical. If \code{TRUE}, the new barrier algorithm is used.
Defaults to \code{TRUE}.}

\item{all_basins}{A \code{SpatRaster} representing hydrological basins.
Required only if \code{new_barrier = FALSE}.}

\item{source_offset}{Numeric. The vertical offset of the source above the
ground in meters. Defaults to \code{0.34}.}

\item{receiver_offset}{Numeric. The vertical offset of the receiver above
the ground in meters. Defaults to \code{1}.}
}
\value{
A list containing:
  \itemize{
    \item \code{noise_propagation}: A list of noise propagation patterns for
      each frequency.
    \item \code{loss}: A list of losses, including:
      \itemize{
        \item \code{ssl}: Spherical spreading loss.
        \item \code{aal}: Atmospheric absorption loss.
        \item \code{veg}: Vegetation and ground cover loss (if
          \code{new_barrier = FALSE}).
        \item \code{wind}: Wind loss.
        \item \code{bar}: Barrier loss.
        \item \code{barwind}: Combined barrier and wind loss (if
          \code{new_barrier = TRUE}).
        \item \code{topo_zones}: Topographic zones (if \code{new_barrier = FALSE}).
      }
  }
}
\description{
This function computes noise propagation patterns considering atmospheric,
geographic, and barrier effects, with options for using a new or old
barrier approach.
}
\examples{
# Example usage of soundprog_one
# Crear un objeto spread.grid
dominio <- spread_grid(xorig = 298811, yorig = 4106479, nx = 100,
                       ny = 100, dgrid = 25, epsg = 26911)

# Crear fuentes como ejemplo
library(sf)
library(dplyr)
library(terra)
fuente <- data.frame("ids" = c("sr1"),
                      "geometry" = c("POINT(300158.3 4107751.4)")) |>
  st_as_sf(wkt = c("geometry"), crs = 26911)

# Datos geográficos
path_dem <- system.file("extdata", "dem.tif", package="rspread")
terreno <- terra::rast(path_dem)/3.2808 # Conversión de pies a metros
path_land <- system.file("extdata", "land.tif", package="rspread")
cobertura <- terra::rast(path_land)
geodatos <- spreadgeo(dominio, terreno, cobertura)

# Cargar frecuencias
path_freq <- system.file("extdata", "all_freqs.csv", package="rspread")
frecuencias <- read.csv(path_freq)
frecuencias <- split(frecuencias, ~ids)[[1]]

results <- soundprog_one(
  sgrid = dominio,
  src_one = fuente,
  data_one = frecuencias,
  atmos = list("temp"=28, "rh"=50, "wd"=105, "ws"=12.8, "seas_cond"=5),
  geo = geodatos,
  new_barrier = TRUE
)

}
