% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread-grid-class.R
\docType{class}
\name{spread.grid-class}
\alias{spread.grid-class}
\title{spread.grid: Clase para representar una grilla espacial}
\description{
spread.grid: Clase para representar una grilla espacial
}
\section{Slots}{

\describe{
\item{\code{xorig}}{Coordenada X de origen.}

\item{\code{yorig}}{Coordenada Y de origen.}

\item{\code{nx}}{Número de celdas en el eje X.}

\item{\code{ny}}{Número de celdas en el eje Y.}

\item{\code{dgrid}}{Tamaño de las celdas.}

\item{\code{epsg}}{Código EPSG del sistema de referencia.}

\item{\code{total_points}}{Número total de puntos.}
}}

