% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{weighting}
\alias{weighting}
\title{Calcula la ponderación de los niveles de presión sonora}
\usage{
weighting(rsl, pond = "A", all_rsl = F)
}
\arguments{
\item{rsl}{Una lista de resultados con niveles de presión sonora en bandas de frecuencia específicas.
Debe contener los elementos `f63`, `f125`, `f250`, `f500`, `f1000`, `f2000`, `f4000` y `f8000`.}

\item{pond}{Un carácter que indica el tipo de ponderación a aplicar. Puede ser:
- "A": Ponderación A, que simula la respuesta del oído humano a niveles bajos.
- "C": Ponderación C, que tiene una respuesta más plana, utilizada en niveles altos.
- "Z": Ponderación Z (plana), sin modificaciones en la respuesta de frecuencia.}

\item{all_rsl}{Un valor lógico. Si es `TRUE`, se espera que `rsl` contenga dos listas de resultados
y se usará la lista `noise_propagation`. Si es `FALSE`, se espera que `rsl` contenga ocho listas de resultados.}
}
\value{
Un valor numérico que representa el nivel de presión sonora ponderado.
}
\description{
Esta función aplica diferentes ponderaciones (A, C o Z) a los niveles de presión sonora
en distintas bandas de frecuencia, basándose en los resultados proporcionados.
}
\examples{
rsl <- list(f63 = 50, f125 = 55, f250 = 60, f500 = 65, f1000 = 70, f2000 = 75, f4000 = 80, f8000 = 85)
weighting(rsl, pond = "A")

}
